requirements.txt
streamlit==1.35.0
tensorflow-cpu==2.15.0
numpy
pillow